+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Folding@home Web Control Changelog
  - Show logged credits in WU view.
  - Fixed RCG display before WU download.  #165
  - Change direct client address hotkey to CTRL+SHIFT+L.  #166
+ - Added direct address validation.
 
 ## v8.4.1
  - Fix ``Finishing`` display bug.